Make XYZ::new take arrays instead of slices
Previously if a slice of len<3 was passed in, the constructor would
panic. And if a slice of len>3 was passed in, all elements after the
three first would just be silently ignored. This seems lossy and error
prone. If requirements can be encoded in the type system (without
introducing a lot of unergonomic APIs), then they should. XYZ::new is
now truly infallible and it will use all the passed in data.