Remove Deref and DerefMut implementation for Colorant
If the Colorant is supposed to stay in the value range [0.0, 1.0],
then implementing DerefMut<Target=Spectrum> violates that rule,
since it's possible to modify the underlying spectrum so the range
requirement is no longer upheld.

The documentation for the Deref trait has recommendations when to and
when not to implement Deref. I do not think Colorant->Spectrum have the
properties fitting this implementation. I instead think a user should
explicitly call the .spectrum() method to get the underlying spectrum.